undo updating numpy op

+++ keras_core/layers/preprocessing/discretization_test.py
@@ -70,7 +70,9 @@ class DicretizationTest(testing.TestCase):
 
     def test_tf_data_compatibility(self):
         # With fixed bins
-        layer = layers.Discretization(bin_boundaries=[0.0, 0.35, 0.5, 1.0])
+        layer = layers.Discretization(
+            bin_boundaries=[0.0, 0.35, 0.5, 1.0], dtype="float32"
+        )
         x = np.array([[-1.0, 0.0, 0.1, 0.2, 0.4, 0.5, 1.0, 1.2, 0.98]])
         self.assertAllClose(layer(x), np.array([[0, 1, 1, 1, 2, 3, 4, 4, 3]]))
         ds = tf_data.Dataset.from_tensor_slices(x).batch(1).map(layer)
